Refactor protocol types and ctl handlers for consistency
- Move JitterState enum to protocol.rs with Serialize/Deserialize,
  change JitterStatsInfo.state from u8 to JitterState
- Add Display impls for ChannelId, UiLoopbackMode, NetLoopbackMode,
  JitterState, and FlashPhase
- Add ChannelId::ALL constant and ChannelConfig::relay_url_display()
- Add helper methods to MgmtBootloaderInfo (version_major/minor, sp,
  reset_handler, sp_valid, reset_valid)
- Add CtlError::is_timeout() and escape_non_ascii() utility
- Move set_sframe_key length validation from ctl handler into CtlCore
- Centralize timeout constants (BOOTLOADER_PROBE_MS, MONITOR_MS,
  NORMAL_SECS) replacing hardcoded values
- Extract enter_mgmt_bootloader() to deduplicate info/flash handlers
- Add interpret_esp32_security() for NET chip security fuse display
- Use uart_config::HIGH_SPEED.baudrate instead of hardcoded 1000000

+++ net/src/main.rs
@@ -1290,7 +1290,7 @@ fn handle_mgmt_message(
                     underruns: s.underruns as u32,
                     overruns: s.overruns as u32,
                     level: s.level as u16,
-                    state: s.state as u8,
+                    state: s.state,
                 };
                 let mut buf = [0u8; 64];
                 if let Ok(serialized) = postcard::to_slice(&info, &mut buf) {
